changes to applying files
